+++ actions/credit-get.go
@@ -6,21 +6,18 @@ import (
 	"github.com/apmath-web/credit-go/viewModels"
 	"log"
 	"net/http"
-	"strconv"
-	"strings"
 )
 
 func Get(response http.ResponseWriter, request *http.Request) {
 	response.Header().Add("Content-Type",
 		"application/json; charset=utf-8")
-	paths := strings.Split(request.URL.Path, "/credit/")
-	id, err := strconv.ParseInt(paths[1], 10, 64)
+	id, err := getId(request.URL.Path)
 	if err != nil {
 		errorMessage("Invalid id format", 400, response)
 		return
 	}
 	repo := Repository
-	credit := repo.Get(int(id))
+	credit := repo.Get(id)
 	if credit == nil {
 		errorMessage("Credit not found", 404, response)
 		return
