+++ actions/payment-put.go
@@ -5,15 +5,12 @@ import (
 	"github.com/apmath-web/credit-go/valueObjects"
 	"github.com/apmath-web/credit-go/viewModels"
 	"net/http"
-	"strconv"
-	"strings"
 )
 
 func PaymentWriteOf(response http.ResponseWriter, request *http.Request) {
 	response.Header().Add("Content-Type",
 		"application/json; charset=utf-8")
-	paths := strings.Split(request.URL.Path, "/credit/")
-	id, err := strconv.ParseInt(paths[1], 10, 64)
+	id, err := getId(request.URL.Path)
 	if err != nil {
 		errorMessage("Invalid id format", 400, response)
 		return
